Rebuild 1.1. Full build of one of the libs was skipped
Seems that you can't rely on `lerna run build` to build everything if
build of one of the packages is failing (even though there's no
dependencies between the packages or dependency is even in the other
direction). Or it's some kind of other issue... But it's related to
cache.

Added a script to do a totally clean build, ignoring cache.